minor refactoring and cleanup in permission handling across auth and oauth2 endpoints.
