+++ pauldron/controllers/oauth2-endpoint.js
@@ -1,7 +1,8 @@
 const _ = require("lodash");
-const hash = require("object-hash");
 const db = require("../lib/db");
 const logger = require ("../lib/logger");
+
+const {reconcilePermissionsAndObligations} = require("../lib/permission-handler");
 
 const TimeStampedPermission = require("../model/TimeStampedPermission");
 const JWTClaimsToken = require("../lib/jwt-claims-token");
@@ -10,8 +11,6 @@ const GenericErrorHandler = require("./error-handler");
 
 const {SimplePolicyDecisionCombinerEngine} = require ("pauldron-policy");
 const {policyTypeToEnginesMap} = require("./policy-endpoint");
-
-const DENY_SCOPES_OBLIGATION_ID = "DENY_SCOPES";
 
 const JWT_BEARER_CLIENT_ASSERTION_TYPE= "urn:ietf:params:oauth:client-assertion-type:jwt-bearer";
 
@@ -54,6 +53,7 @@ async function checkPolicies(claims, permissions, policies) {
     if (decision.authorization === "Permit") {
         const grantedPermissions = reconcilePermissionsAndObligations(permissions, decision.obligations);
         if (!grantedPermissions || !grantedPermissions.length) {
+            logger.debug("Rejecting token request because to permissions could be granted.");
             throw {
                 error: "policy_forbidden"
             };
@@ -64,33 +64,6 @@ async function checkPolicies(claims, permissions, policies) {
             error: "policy_forbidden"
         }; //fail safe for anything other than an explicit permit
     }
-}
-
-function reconcilePermissionsAndObligations (permissions, obligations) {
-    const deniedScopes = obligations[DENY_SCOPES_OBLIGATION_ID];
-
-    if (deniedScopes) {
-        const grantedPermissions = permissions.map(
-            (permission) => (
-              {
-                  resource_set_id: permission.resource_set_id,
-                  scopes: (permission.scopes || []).filter(
-                      (scope) => (! arrayDeepIncludes (deniedScopes, scope))
-                  )
-              }
-        ));
-
-      return grantedPermissions.filter((permission) => (
-        ((permission.scopes.length || 0) !== 0)
-      ));
-    } else {
-      return permissions;
-    }
-}
-
-function arrayDeepIncludes(array, thing) {
-    const arrayHashes = array.map((element) => (hash(element)));
-    return arrayHashes.includes(hash(thing));
 }
 
 const yup = require("yup");
